#7271 Check missing config file

+++ conans/client/conan_api.py
@@ -604,15 +604,14 @@ class ConanAPIV1(object):
 
     @api_method
     def config_install_list(self):
-        try:
-            return json.loads(load(self.app.cache.config_install_file))
-        except FileNotFoundError:
+        if not os.path.isfile(self.app.cache.config_install_file):
             raise ConanException("There is no config data. Need to install config first.")
-        except Exception as error:
-            raise ConanException("Could not load config data: {}".format(error))
+        return json.loads(load(self.app.cache.config_install_file))
 
     @api_method
     def config_install_remove(self, index):
+        if not os.path.isfile(self.app.cache.config_install_file):
+            raise ConanException("There is no config data. Need to install config first.")
         configs = json.loads(load(self.app.cache.config_install_file))
         try:
             configs.pop(index)
